Fix incorrect queuetime in hit payload urls

+++ gai++/src/DataStore.h
@@ -198,7 +198,7 @@ namespace GAI
         /// @return
         ///     Generated Hit instance
         ///
-        Hit createHit(const std::string& version, const std::string& url, const double timestamp)
+        Hit createHit(const std::string& version, const std::string& url, const uint64_t timestamp)
         {
             return Hit( version, url, timestamp);
         }
